Refactored env.js to work with Expo release channels on the front end
